Converted `unittest`-style tests to `pytest`-style tests (#16)
I want to use the [fixture features](https://docs.pytest.org/en/stable/how-to/fixtures.html) that
`pytest` provides, and it's quite hard to do with the `unittest`-style
tests, so I converted all the tests to use the `pytest` style. I think
it's a bit more readable as well.