Implement `ProxyConfiguration` class and `Actor.create_proxy_configuration()` (#18)
This implements the `ProxyConfiguration` class and
`Actor.create_proxy_configuration()`, modelled after the JavaScript SDK.

I used Pydantic for the argument validation, but I'm not really happy
with it, it has a few bugs and weird behaviors. I might change it later
for some other argument validation library, if I can find something as
good as `ow` for TypeScript.

The unit tests were the biggest work, I learned a lot about `pytest`
while implementing them.

I'm also not sure about the custom monkeypatcher for `ApifyClient`, I
made it kinda general purpose so that we can reuse it in other unit
tests, but perhaps it would be better to mock the Apify API directly
using `respx`.