Add logger (#69)
This PR adds a new submodule, `apify.log`, and in it a `logger`, which
is just a standard logger from the `logging` library, with name `apify`
as per best practices for Python logging. It also makes it available
under the `Actor.log` property, for convenience.

The main magic happens in the `ActorLogFormatter`, which is a log
formatter that consumes the logging messages, and formats them so they
are pretty:

<img width="1327" alt="Screenshot 2023-02-15 at 10 31 01"
src="https://user-images.githubusercontent.com/2934111/218988666-97fdc1b3-c33f-4f71-a7f1-842a86de8433.png">

The log formatter and handler is attached in the actor setup, so in the
`__main__.py` file. We'll have to put the setup in all the Python actor
templates so that it's applied in the actors by default.

The integration tests and unit tests might be too strict, and any slight
change might break them, so if it happens too often, we can rewrite them
a bit.