Fix handling the `MIGRATING` actor event (#70)
The handler for the `MIGRATING` event was missing the argument for event
data, so it was broken. This was very hard to discover because you can't
force actor migrations easily.

This fixes it, plus adds a unit test for checking this behavior.

It also makes sure that the final `PERSIST_STATE` event is not emitted
twice.