`MemoryStorage` improvements (#78)
There was one problem in `MemoryStorage` that I wanted to fix: if you
called `Actor.open_dataset(id='someId')`, if the dataset with ID
`someId` did not exist, it instead created a new named dataset with name
`someId`.

I thought this would be a simple fix, but in the end this was harder
than expected, and while working on it I decided to refactor the
`MemoryStorage` a bit to make it cleaner. It diverges from the Crawlee
implementation more, but I believe in this case it does not matter, and
on the outside it should still behave the same.

I did:
- split all the `id_or_name` parameters into separate `id` and `name`
- made `Dataset`, `KeyValueStore` and `RequestQueue` subclasses of
`BaseStorage`
- removed `StorageManager` and moved the storage caching logic to
`BaseStorage`
- there are now three separate caches, one for each storage type
	- they each live in its corresponding storage class
- the caching implementation is shared in `BaseStorage`, with only tiny
bits overriden in each storage class
- no need to pass the storage classes around anymore! and typing works
nicely with just one `cast`
- in `MemoryStorage`:
	- I changed the logic of opening storages by ID
- if the resource client is not cached, it tries to find the storage by
looking into the `storages/<STORAGE_TYPE>` directory, reading the
metadata of every storage, and finding the right storage by ID in the
metadata
- I made all the resource clients subclass `BaseResourceClient`, where
the caching implementation lives now, with only small overrides in each
class
- I changed all the generated IDs to be Apify-like
`crypto_random_object_id`s, instead of UUIDs
- I added tests for:
	- the `id` and `name` split, 
	- `Actor.open_<STORAGE_TYPE>(..., force_cloud=True)`

There are still some issues:
- when you do `DatasetCollectionClient.list()`, it only returns storages
that were already open by the actor (this applies to all storages)
- the caching logic in the resource clients is not very clean, it's
weird how one instance of the resource client internally uses another
instance of the same class, I think the caching logic should be more
like in the `BaseStorage`

Sorry for such a massive rewrite, I didn't think it would be so big when
I started with it.