Add option to add event handlers which accept no arguments (#84)
Right now, you can add actor event handlers which accept exactly one
argument - event data - even if the event itself does not convey any
data (like the migration event). If you try to pass an event handler
which accepts no arguments, the event handler will throw when the event
happens.

Meaning that this does not work:
```python
def migration_handler():
    print('Migration happened')

Actor.on(ActorEventTypes.MIGRATING, migration_handler)
```

But this works, because the `MIGRATING` event never has any data:
```python
def migration_handler(event_data: None):
    assert event_data is None
    print('Migration happened')

Actor.on(ActorEventTypes.MIGRATING, migration_handler)
```

One user was struggling with this, and it's pretty hard to test,
especially for the migrations, so I changed the logic so that the first
thing works, and the event handler is passed the `event_data` argument
only if it can accept it.