Fix detecting signature of built-in functions in event handler (#86)
I found out that `inspect.signature()` cannot read the signature of
built-in methods, so e.g. `Actor.on(ActorEventTypes.SYSTEM_INFO, print)`
would not work.

This fixes it by rewriting the signature-processing to ignore errors
when the signature can't be read, and also to better check methods which
have default arguments, by just trying to bind the arguments to the
method signature.