Start enforcing local storage to always use the UTF-8 encoding (#87)
This fixes a bug reported on Discord, where a user on Windows had issues
with datasets loaded from the `storage` directory not being able to
parse some items which had special characters, because the items were
saved with the `utf-8` encoding, but loaded with the default Windows
encoding `cp-1252`. This forces `utf-8` to be used everywhere.