feat: add signing of public URL (#407)
This PR is part of [Issue
#19363](https://github.com/apify/apify-core/issues/19363), which updates
`KeyValueStore.getPublicUrl(recordKey)` to generate signed links using
HMAC.

**This PR**:
- creates signature and appends it to the public URL of the record in KV
store

P.S. Before merging, we need to wait for the release of Crawlee, so we
can update version of Crawlee.

P.P.S It's hard to test the changes, since master branch of SDK and
master branch of Crawlee are out of sync. Crawlee has some breaking
changes in version 6.0, which are not yet addressed in master branch of
SDK.

Previous PR that adds storageObject to Crawlee Python is
[here](https://github.com/apify/crawlee-python/pull/993)

Same PR in SDK JS is
[here](https://github.com/apify/apify-sdk-js/pull/358)