feat: Add `RemainingTime` option for `timeout` argument of `Actor.call` and `Actor.start` (#473)
### Description

Added convenient way to start another Actor with reasonable timeout to
prevent lingering actor runs after main actor shuts down.
Added test.

`RemainingTime` value of the timeout argument is not a default one and
has to be explicitly passed by the user. Using this value will calculate
remaining time of this actor run and pass it as a timeout to the actor
that is being started.

### Issues

- Closes: #472

---------

Co-authored-by: Vlada Dusek <v.dusek96@gmail.com>